background/requirements.txt
SQLAlchemy==2.0.28
redis==5.0.3
python-dotenv==1.0.1